refactor : 문서역사 데이터 파싱 함수 분리(#115)
